requirements.txt
# hokuriku-tourism-ai-governance
# Python 3.10+
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
statsmodels>=0.14
scipy>=1.11
jpholiday>=0.1.8
PyYAML>=6.0
japanize-matplotlib>=1.1.3
